fix ling int bug for scanner

+++ src/com/apps4better/recycle4better/view/MainActivity.java
@@ -141,7 +141,7 @@ public void startScan (){
 public void onActivityResult(int requestCode, int resultCode, Intent intent) {
 	  IntentResult scanResult = IntentIntegrator.parseActivityResult(requestCode, resultCode, intent);
 	  if (scanResult != null) {
-			  long id = Integer.valueOf(scanResult.getContents()).intValue();
+			  long id = Long.valueOf(scanResult.getContents()).longValue();
 				Intent i = new Intent(context, ProductDetailActivity.class);
 				i.putExtra("product_id", id);
 				i.putExtra("load_info", true);
